+++ Source/Scene/Globe.js
@@ -250,6 +250,14 @@ import TileSelectionResult from './TileSelectionResult.js';
          */
         this.atmosphereBrightnessShift = 0.0;
 
+        /**
+         * Whether to cull back-facing terrain. Set this to false when viewing terrain from below the surface.
+         *
+         * @type {Boolean}
+         * @default true
+         */
+        this.backFaceCulling = true;
+
         this._oceanNormalMap = undefined;
         this._zoomedOutOceanSpecularIntensity = undefined;
     }
@@ -739,7 +747,7 @@ import TileSelectionResult from './TileSelectionResult.js';
             tileProvider.saturationShift = this.atmosphereSaturationShift;
             tileProvider.brightnessShift = this.atmosphereBrightnessShift;
             tileProvider.fillHighlightColor = this.fillHighlightColor;
-
+            tileProvider.backFaceCulling = this.backFaceCulling;
             surface.beginFrame(frameState);
         }
     };